fix and operator

+++ app/src/main/java/com/baijunty/printer/EncryptUtil.kt
@@ -1,6 +1,5 @@
 package com.baijunty.printer
 
-import okhttp3.internal.and
 import java.security.MessageDigest
 
 object EncryptUtil {
@@ -10,7 +9,7 @@ object EncryptUtil {
         val b = md.digest()
         val buf = StringBuilder()
         for (k in b) {
-            val i = k and 0xff
+            val i = k.toInt() and 0xff
             if (i < 16) buf.append("0")
             buf.append(Integer.toHexString(i))
         }
